refactor(ui): update resolver imports
